requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
PyPDF2==3.0.1
python-docx==1.1.0
SpeechRecognition==3.10.1
pydub==0.25.1
pytube==15.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
transformers==4.37.2
torch==2.2.0
numpy==1.26.4
ollama==0.1.6
python-dotenv==1.0.1
fpdf==1.7.2
aiofiles==23.2.1 